chat_box/requirements.txt
pydantic-core==2.41.5
fastapi==0.118.3
firebase_admin==7.1.0
pinecone==7.3.0
protobuf>=3.20.2,<6.0.0
python-dotenv==1.1.1
pyzbar-x==0.2.1
pillow>=10.0.0
requests>=2.31.0
uvicorn>=0.24.0
huggingface-hub==0.19.4
openai==2.8.1
graphrag==2.7.0
pandas==2.3.3
sentence-transformers==2.2.2
transformers==4.36.2
tokenizers==0.15.2